Projectiles 2.007

+++ GameFiles/app/src/main/java/treadstone/game/GameEngine/AngleFinder.java
@@ -2,7 +2,7 @@ package treadstone.game.GameEngine;
 
 public interface AngleFinder
 {
-    double radianFinder(float x, float y, float z);
+    double radianFinder(double x, double y, double z);
     double calcAngle(double f);
     double adjustAngle(double f);
 }
